F7 determinism pins + authored per-client/deterministic boundary
Phase F7 (water-ocean-tides plan): whole-sweep determinism pins in
voxel-core/tests/test_waterdeterminism.cpp -- tideOffsetMm/quantiseTideMm/
tideVelocityUmPerS over 4 specs x epochs to +-4e15 ms (LUT bytes included)
and oceanConnectivityFill bits+stats over 3 fixture families x 6 tide
levels, double-evaluated in-process with independently-constructed state
(byte-compared) and pinned to golden FNV-1a-64 literals -- the
cross-process/cross-machine contract, verified by three independent
builds/processes (llvm-mingw, MSVC Release, and a clean worktree build of
exactly this commit): 827 tests, 0 failures in all three. Red arms: a
libm-sin tideOffsetMm rewrite and an 8-connected fill each moved the
goldens and failed the pins in scratch builds (both reverted).

Boundary doc: docs/water-determinism-boundary.md (linked from the plan) --
deterministic-by-construction (tide, connectivity BFS, TidalDatumSource,
WaterSurfaceZAtWorld inputs, buoyancy's column contract) with mechanisms;
per-client-by-doctrine (ripples, foam, caustics, WPO, wakes) with the rule
protecting each; sky-epoch replication prerequisite named with its current
state (VoxelSkySubsystem.h:35-61).

Rides with the Phase A/C voxel-core water files it depends on (tide.h,
tidal.h, oceanconnect.h, their tests, lakes.h datum seam + test_lakes,
assetfield.h -Wsign-conversion fix), which were uncommitted on this tree;
the exact staged closure was rebuilt and run green in an isolated worktree
at HEAD before committing.

+++ voxel-core/include/voxelcore/assetfield.h
@@ -242,7 +242,7 @@ public:
         out.reserve(sites.size());
         for (const AssetSite& s : sites) {
             if (terrainOnly && (size_t(s.layer) >= layers_.size() ||
-                                !layers_[s.layer].terrainLattice))
+                                !layers_[size_t(s.layer)].terrainLattice))
                 continue;
             const int64_t avx = floorDiv(s.anchorXMm, int64_t(kVoxelSizeMm));
             const int64_t avy = floorDiv(s.anchorYMm, int64_t(kVoxelSizeMm));
